[bugfix/markdown-and-text-changes] create an extension for the attributed string
